requirements.txt
Django==2.2.19
djangorestframework==3.13.1
flake8==4.0.1
mccabe==0.6.1
numpy==1.23.0
pandas==1.4.3
pycodestyle==2.8.0
pyflakes==2.4.0
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
